requirements.txt
requests
argparse
dateutil
pytest
pyyaml
yattag